requirements.txt
asgiref==3.5.0
click==8.0.3
Django==3.2.12
djangorestframework==3.13.1
Faker==12.1.0
importlib-metadata==4.10.1
joblib==1.1.0
lxml==4.7.1
nltk==3.6.7
numpy==1.21.5
pdfminer==20191125
Pillow==9.0.0
pycryptodome==3.13.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
regex==2022.1.18
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
typing-extensions==4.0.1
zipp==3.7.0
